Move embed player to mp3 stream (#88)
* Move embed player to mp3 stream

* remove logs
